products crud checkup fixes

+++ src/app/admin/products/actions.ts
@@ -2,24 +2,24 @@
 
 import db from "@/drizzle";
 import { product } from "@/drizzle/schema";
-import { executeAction } from "@/drizzle/utils/executeAction";
+import { executeAction, type OperationResult } from "@/drizzle/utils/executeAction";
 import { eq } from "drizzle-orm";
 import fs from "fs/promises";
 import { revalidatePath } from "next/cache";
 
-export const deleteProduct = async (id: number) => {
+export const deleteProduct = async (state: OperationResult | undefined, id: number) => {
     return executeAction({
         actionFn: async () => {
             const thisProduct = await db.query.product.findFirst({
                 columns: {
                     thumbnail: true,
                 },
-                where: eq(product.id, id),
+                where: eq(product.id, +id),
             });
-
-            await fs.unlink(`public${thisProduct?.thumbnail}`);
-            await db.delete(product).where(eq(product.id, id));
-
+            if (!!thisProduct?.thumbnail) {
+                await fs.unlink(`public${thisProduct?.thumbnail}`);
+            }
+            await db.delete(product).where(eq(product.id, +id));
             revalidatePath("/admin/products");
             revalidatePath("/");
             revalidatePath("/products");
